refactor: extract class/race normalizers to dndBeyond-classes.ts
Extracted normalizeClasses, normalizeClassEntry, normalizeRace from
lib/dndBeyondCharacterImport.ts into lib/import/dndBeyond-classes.ts.

Also exported isPresent and createValidationError from dndBeyond-utils.ts
to support the new module.

No behavior change - pure structural refactor.
Closes #151
