+++ openspec/changes/archive/2026-05-03-extract-dnd-ability-scores-foundation/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: sdd-with-feedback-loop
+created: 2026-05-02
